fix: unmount tab in App
When selecting another tab, the current one is unmounted, making it totally gone from the DOM
and clear all ongoing shit inside that tab. For example, before this commit, the clocks is
just hidden when switching to other tab, and `tick()` is still called periodically.

Refactor `ClockPane`. Now it's weirder...